+++ src/components/app/userActionFormVotingInformationResearched/constants.ts
@@ -1,0 +1,7 @@
+export const ANALYTICS_NAME_USER_ACTION_FORM_VOTING_INFORMATION_RESEARCHED =
+  'User Action Form Voting Information Researched'
+
+export enum SECTIONS_NAMES {
+  ADDRESS = 'Address',
+  SUCCESS = 'Success',
+}
