+++ src/components/app/userActionFormVotingInformationResearched/sections/success.tsx
@@ -1,0 +1,10 @@
+import { USER_ACTION_FORM_SUCCESS_SCREEN_INFO } from '@/components/app/userActionFormSuccessScreen/constants'
+import { UserActionFormSuccessScreenFeedback } from '@/components/app/userActionFormSuccessScreen/UserActionFormSuccessScreenFeedback'
+
+export const UserActionFormVotingInformationResearchedSuccess = () => {
+  return (
+    <UserActionFormSuccessScreenFeedback
+      {...USER_ACTION_FORM_SUCCESS_SCREEN_INFO.VOTING_INFORMATION_RESEARCHED}
+    />
+  )
+}
